Refactor TcpSocketDataSender to a plugin form
