+++ osu!StreamCompanion.sln
@@ -35,6 +35,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BeatmapPpReplacements", "pl
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FileMapDataSender", "plugins\FileMapDataSender\FileMapDataSender.csproj", "{32A61CD7-FA30-40CB-8A23-788BBD65D4A7}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TcpSocketDataSender", "plugins\TcpSocketDataSender\TcpSocketDataSender.csproj", "{50E3D207-3812-433A-8F9A-757FF55C991A}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -211,6 +213,18 @@ Global
 		{32A61CD7-FA30-40CB-8A23-788BBD65D4A7}.Remote debug|Any CPU.Build.0 = Release|Any CPU
 		{32A61CD7-FA30-40CB-8A23-788BBD65D4A7}.Remote debug|x86.ActiveCfg = Release|Any CPU
 		{32A61CD7-FA30-40CB-8A23-788BBD65D4A7}.Remote debug|x86.Build.0 = Release|Any CPU
+		{50E3D207-3812-433A-8F9A-757FF55C991A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{50E3D207-3812-433A-8F9A-757FF55C991A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{50E3D207-3812-433A-8F9A-757FF55C991A}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{50E3D207-3812-433A-8F9A-757FF55C991A}.Debug|x86.Build.0 = Debug|Any CPU
+		{50E3D207-3812-433A-8F9A-757FF55C991A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{50E3D207-3812-433A-8F9A-757FF55C991A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{50E3D207-3812-433A-8F9A-757FF55C991A}.Release|x86.ActiveCfg = Release|Any CPU
+		{50E3D207-3812-433A-8F9A-757FF55C991A}.Release|x86.Build.0 = Release|Any CPU
+		{50E3D207-3812-433A-8F9A-757FF55C991A}.Remote debug|Any CPU.ActiveCfg = Release|Any CPU
+		{50E3D207-3812-433A-8F9A-757FF55C991A}.Remote debug|Any CPU.Build.0 = Release|Any CPU
+		{50E3D207-3812-433A-8F9A-757FF55C991A}.Remote debug|x86.ActiveCfg = Release|Any CPU
+		{50E3D207-3812-433A-8F9A-757FF55C991A}.Remote debug|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -226,6 +240,7 @@ Global
 		{16BB58FA-897B-4A75-BEDC-D42428EB8312} = {7357CB17-E9D7-493F-BC1B-354F717D7AA3}
 		{0C50FF12-C298-424F-8B56-E65BCDDAB159} = {7357CB17-E9D7-493F-BC1B-354F717D7AA3}
 		{32A61CD7-FA30-40CB-8A23-788BBD65D4A7} = {7357CB17-E9D7-493F-BC1B-354F717D7AA3}
+		{50E3D207-3812-433A-8F9A-757FF55C991A} = {7357CB17-E9D7-493F-BC1B-354F717D7AA3}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {90CD718D-2503-4BBC-867D-E8EF9C9DBC77}
