+++ osu!StreamCompanion/packages.config
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="EntityFramework" version="6.0.0" targetFramework="net4-client" />
-  <package id="Newtonsoft.Json" version="10.0.3" targetFramework="net40-client" />
+  <package id="Newtonsoft.Json" version="11.0.2" targetFramework="net40-client" />
   <package id="System.Data.SQLite" version="1.0.98.1" targetFramework="net4-client" />
   <package id="System.Data.SQLite.Core" version="1.0.98.1" targetFramework="net4-client" />
   <package id="System.Data.SQLite.EF6" version="1.0.98.1" targetFramework="net4-client" />
