+++ plugins/TcpSocketDataSender/TcpSocketDataGetter.cs
@@ -2,14 +2,11 @@
 using System.Collections.Generic;
 using System.Windows.Forms;
 using Newtonsoft.Json;
-using osu_StreamCompanion.Code.Core;
-using osu_StreamCompanion.Code.Helpers;
-using osu_StreamCompanion.Code.Misc;
 using StreamCompanionTypes;
 using StreamCompanionTypes.DataTypes;
 using StreamCompanionTypes.Interfaces;
 
-namespace osu_StreamCompanion.Code.Modules.MapDataGetters.TcpSocket
+namespace TcpSocketDataSender
 {
     public class TcpSocketDataGetter : IModule, IMapDataGetter, ISettingsProvider, IDisposable
     {
