+++ plugins/TcpSocketDataSender/TcpSocketSettings.resx
